requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
psycopg[binary,pool]==3.2.3
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
